cambios varios

+++ public/formulacion/modulos/reportes/acFormato.php
@@ -73,7 +73,7 @@ class MYPDF extends TCPDF {
 		left join t45_planes_zulia as t45 on t49.co_area_estrategica=t45.co_area_estrategica and t45.nu_nivel = 0
 		left join t45_planes_zulia as t45a on t49.co_area_estrategica=t45a.co_area_estrategica and t49.co_ambito_estado=t45a.co_ambito_zulia and t45a.nu_nivel = 1
 		left join t45_planes_zulia as t45b on t49.co_ambito_estado=t45b.co_ambito_zulia and t49.co_macroproblema=t45b.co_macroproblema and t45b.nu_nivel = 3 and t45b.edo_reg = true
-                left join t45_planes_zulia as t45c on t49.co_ambito_estado=t45c.co_ambito_zulia and t49.co_nodos=t45c.co_nodo and t45c.nu_nivel = 4 and t45c.edo_reg = true
+                left join t45_planes_zulia as t45c on t49.co_ambito_estado=t45c.co_ambito_zulia and t49.co_nodos::integer=t45c.co_nodo and t45c.nu_nivel = 4 and t45c.edo_reg = true
 	where t46.edo_reg is true and ".$condicionAC." t47.edo_reg is true AND t46.id_ejercicio = ".$_SESSION['ejercicio_fiscal']." order by 9, 8, 1, 17 ASC";
 
 		$this->datos = $comunes->ObtenerFilasBySqlSelect($sql);
